Started sell product detail

+++ imports/ui/models/inventories/new/InventoriesNew.jsx
@@ -62,7 +62,7 @@ export default class InventoriesNew extends React.Component{
               defaultDate={new Date()}
               formatDate={new DateTimeFormat('en-US', {
                 day: 'numeric',
-                month: 'numeric',
+                month: 'short',
                 year: 'numeric',
               }).format} />
           </div>
@@ -75,7 +75,7 @@ export default class InventoriesNew extends React.Component{
               fullWidth={true}
               formatDate={new DateTimeFormat('en-US', {
                 day: 'numeric',
-                month: 'numeric',
+                month: 'short',
                 year: 'numeric',
               }).format} />
           </div>
